+++ device-minion-runtime/src/MasterMinion/include/services/host_cmd_hdlr.h
@@ -34,8 +34,8 @@
 typedef struct execution_cycles_ {
     uint64_t cmd_start_cycles;
     uint64_t exec_start_cycles;
+    uint32_t exec_prev_cycles;
     uint32_t wait_cycles;
-    uint8_t pad[4];
 } __attribute__((packed, aligned(8))) execution_cycles_t;
 
 /*! \fn int32_t Host_Command_Handler(void* command_buffer, uint8_t sqw_idx,