+++ device-minion-runtime/src/MasterMinion/include/workers/kw.h
@@ -112,4 +112,10 @@ int32_t KW_Dispatch_Kernel_Abort_Cmd(
 */
 void KW_Abort_All_Dispatched_Kernels(uint8_t sqw_idx);
 
+/*! \fn uint64_t KW_Get_Average_Exec_Cycles(void)
+    \brief This function gets Compute Minion utlization.
+    \return Average consumed cycles
+*/
+uint64_t KW_Get_Average_Exec_Cycles(void);
+
 #endif /* KW_DEFS_H */