+++ device-minion-runtime/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [[_TOC_]]
 ## [Unreleased]
 ### Added
+- [SW-13352] DMA and Kernel utilization stats support.
 ### Changed
 ### Deprecated
 ### Removed